+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [RUSTSEC-2024-0021]: Update eyre from 0.6.11 to 0.6.12 ([@QuantumDancer](https://github.com/QuantumDancer))
 
 ### Added
+- API: The health check (`GET /health_check`) now fails with an `500 INTERNAL SERVER ERROR` if no connection can be made to the database (#622) ([@QuantumDancer](https://github.com/QuantumDancer))
 - CI: Add workflow for testing the pyaudtitor source distribution ([@dirksammel](https://github.com/dirksammel))
 - CI: Add workflow for building the Rust documentation ([@dirksammel](https://github.com/dirksammel))